combined and renamed some files and classes
