+++ geomesa-feature/src/main/scala/org/locationtech/geomesa/feature/kryo/KryoGeometrySerializer.scala
@@ -21,7 +21,7 @@ import java.io.{InputStream, OutputStream}
 import com.esotericsoftware.kryo.io.{Input, Output}
 import com.esotericsoftware.kryo.{Kryo, Serializer}
 import com.vividsolutions.jts.geom.Geometry
-import org.locationtech.geomesa.feature.serialization.kryo.{KryoReader, KryoWriter}
+import org.locationtech.geomesa.feature.serialization.{KryoReader, KryoWriter}
 
 import scala.ref.SoftReference
